Modify post gpio config callback
GPIO callback to board specific function was implemented in the wrong place.
Originally it was defined after gpios has been configured, but this is not good
enough since the MPPs (pin muxing) hasn't been configured yet.

This patch moves the call to the callback function after both GPIOs and MPPs
has been configured.

The bug was found when implementing resetting the on usom phy which uses
GPIO19 (valid on ClearFog boards rev 2.1 and newer).